Move template model `builtin_labels` into `RuntimeModel`
