+++ crates/codegen/runtime/cargo/src/runtime/kinds.rs.jinja2
@@ -49,7 +49,7 @@ impl metaslang_cst::NonterminalKind for NonterminalKind {}
 #[cfg_attr(not(feature = "slang_napi_interfaces"), derive(Clone, Copy))]
 pub enum EdgeLabel {
     // Built-in:
-    {% for label in builtin_labels -%}
+    {% for label in model.builtin_labels -%}
         {{ label | pascal_case }},
     {%- endfor %}
